chore: support nextjs 14.2.0
