+++ src/components/navbar/theme-switcher.tsx
@@ -1,97 +1,28 @@
 "use client"
-import type { MenuOptionItemProps } from "@yamada-ui/react"
-import {
-  Menu,
-  MenuButton,
-  MenuList,
-  MenuOptionItem,
-  MenuOptionGroup,
-  useColorMode,
-  useDisclosure,
-  Box,
-  IconButton,
-  HStack,
-  Text,
-} from "@yamada-ui/react"
+import { useColorMode, Box, IconButton } from "@yamada-ui/react"
 import { memo } from "react"
-import { DisplayIcon, MoonIcon, SunIcon } from "components/icons"
+import { MoonIcon, SunIcon } from "components/icons"
 
 export const ThemeSwitcher = memo(({ locale }: { locale: string }) => {
-  const { changeColorMode, internalColorMode } = useColorMode()
-  const menuOptionItemProps: MenuOptionItemProps = {
-    flexDirection: "row-reverse",
-    justifyContent: "space-between",
-  }
-
-  const { isOpen, onOpen, onClose } = useDisclosure()
+  const { toggleColorMode } = useColorMode()
 
   return (
     <Box>
-      <Menu
-        isOpen={isOpen}
-        onClose={onClose}
-        onOpen={onOpen}
-        closeOnSelect
-        closeOnEsc
-        isLazy
-        lazyBehavior="unmount"
-        placement="bottom"
+      <IconButton
+        type="button"
+        title={locale === "ja" ? "テーマを変える" : "Switch Theme"}
+        rounded="full"
+        onClick={toggleColorMode}
+        _focus={{
+          outline: "3px solid",
+          outlineColor: "hsla(210, 50%, 85%, 0.5)",
+        }}
+        variant="ghost"
+        transitionDuration="300ms"
       >
-        <MenuButton
-          as={IconButton}
-          type="button"
-          title={locale === "ja" ? "テーマを変える" : "Switch Theme"}
-          data-open={isOpen}
-          rounded="full"
-          outline={isOpen ? "3px solid" : "none"}
-          outlineColor="rgba(191, 219, 254, 0.5)"
-          variant="ghost"
-          transitionDuration="300ms"
-          icon={
-            <>
-              <MoonIcon
-                display={["none", "block"]}
-                color="indigo.500"
-                size="2xl"
-              />
-              <SunIcon
-                display={["block", "none"]}
-                color="yellow.500"
-                size="2xl"
-              />
-            </>
-          }
-          // TODO active scale
-        />
-
-        <MenuList>
-          <MenuOptionGroup
-            title="theme"
-            type="radio"
-            onChange={changeColorMode}
-            value={internalColorMode}
-          >
-            <MenuOptionItem {...menuOptionItemProps} value="dark">
-              <HStack as="span" gap="2">
-                <MoonIcon />
-                <Text>Dark</Text>
-              </HStack>
-            </MenuOptionItem>
-            <MenuOptionItem {...menuOptionItemProps} value="light">
-              <HStack as="span" gap="2">
-                <SunIcon />
-                <Text>Light</Text>
-              </HStack>
-            </MenuOptionItem>
-            <MenuOptionItem {...menuOptionItemProps} value="system">
-              <HStack as="span" gap="2">
-                <DisplayIcon />
-                <Text>System</Text>
-              </HStack>
-            </MenuOptionItem>
-          </MenuOptionGroup>
-        </MenuList>
-      </Menu>
+        <MoonIcon display={["none", "block"]} color="indigo.500" size="2xl" />
+        <SunIcon display={["block", "none"]} color="yellow.500" size="2xl" />
+      </IconButton>
     </Box>
   )
 })
